wip: merge cluster and context commands
