+++ src/commands/cluster/configs.ts
@@ -15,17 +15,27 @@
  *
  */
 
-import {Flags as flags} from '../flags.js';
+import {type NodeAlias} from '../../types/aliases.js';
 
-export const USE_FLAGS = {
-  requiredFlags: [],
-  requiredFlagsWithDisabledPrompt: [],
-  optionalFlags: [
-    flags.devMode,
-    flags.quiet,
-    flags.clusterName,
-    flags.context,
-    flags.namespace,
-    flags.userEmailAddress,
-  ],
+export const CONNECT_CONFIGS_NAME = 'connectConfig';
+
+export const connectConfigBuilder = async function (argv, ctx, task) {
+  const config = this.getConfig(CONNECT_CONFIGS_NAME, argv.flags, [
+    'currentDeploymentName',
+  ]) as ClusterConnectConfigClass;
+
+  // set config in the context for later tasks to use
+  ctx.config = config;
+
+  return ctx.config;
 };
+
+export interface ClusterConnectConfigClass {
+  app: string;
+  cacheDir: string;
+  devMode: boolean;
+  namespace: string;
+  nodeAlias: NodeAlias;
+  context: string;
+  clusterName: string;
+}
