+++ src/commands/index.ts
@@ -14,8 +14,7 @@
  * limitations under the License.
  *
  */
-import {ClusterCommand} from './cluster.js';
-import {ContextCommand} from './context/index.js';
+import {ClusterCommand} from './cluster/index.js';
 import {InitCommand} from './init.js';
 import {MirrorNodeCommand} from './mirror_node.js';
 import {NetworkCommand} from './network.js';
@@ -33,7 +32,6 @@ import {type Opts} from '../types/command_types.js';
 export function Initialize(opts: Opts) {
   const initCmd = new InitCommand(opts);
   const clusterCmd = new ClusterCommand(opts);
-  const contextCmd = new ContextCommand(opts);
   const networkCommand = new NetworkCommand(opts);
   const nodeCmd = new NodeCommand(opts);
   const relayCmd = new RelayCommand(opts);
@@ -45,7 +43,6 @@ export function Initialize(opts: Opts) {
     initCmd.getCommandDefinition(),
     accountCmd.getCommandDefinition(),
     clusterCmd.getCommandDefinition(),
-    contextCmd.getCommandDefinition(),
     networkCommand.getCommandDefinition(),
     nodeCmd.getCommandDefinition(),
     relayCmd.getCommandDefinition(),
